Revert "Refactor Resource Commands"
This reverts commit eaaf24fee4b1fe03935255a0e97850ba743ead3a.
and change Delete to use old method
