+++ cli/src/main/java/com/michelin/ns4kafka/cli/GetSubcommand.java
@@ -3,7 +3,11 @@ package com.michelin.ns4kafka.cli;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 
+import javax.inject.Inject;
+
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.michelin.ns4kafka.cli.client.NamespacedResourceClient;
+import com.michelin.ns4kafka.cli.client.NonNamespacedResourceClient;
 import com.michelin.ns4kafka.cli.models.Resource;
 import com.michelin.ns4kafka.cli.models.ResourceDefinition;
 
@@ -11,12 +15,18 @@ import io.micronaut.context.annotation.Value;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.client.exceptions.HttpClientResponseException;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.Help.Ansi;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Parameters;
+import picocli.CommandLine.Help.Ansi;
 
 @Command(name = "get" , description = "Get resources of a Namespace")
-public class GetSubcommand extends AbstractResourceCommand implements Callable<Integer>{
+public class GetSubcommand extends AbstractJWTCommand implements Callable<Integer>{
+
+    @Inject
+    NamespacedResourceClient namespacedClient;
+
+    @Inject
+    NonNamespacedResourceClient nonNamespacedClient;
 
     @Option(names = {"-n", "--namespace"})
     String namespace = "";
